backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pymupdf==1.24.10
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
pillow==10.4.0
langchain==0.3.7
langchain-community==0.3.7
langgraph==0.2.45
chromadb==0.5.18
sentence-transformers==3.2.1
openai==1.54.0
httpx<0.28
python-dotenv==1.0.1
pydantic-settings==2.6.1
langsmith==0.1.140